Trait pallet_signature_bridge::pallet::Config

pub trait Config<I: 'static = ()>: Config {

    type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>;
    type AdminOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type Proposal: Parameter + Dispatchable<RuntimeOrigin = Self::RuntimeOrigin> + EncodeLike + Decode + GetDispatchInfo;
    type SetResourceProposalFilter: Contains<Self::Proposal>;
    type ExecuteProposalFilter: Contains<Self::Proposal>;
    type ChainId: Encode + Decode + Parameter + AtLeast32Bit + Default + Copy + From<u64> + From<u32> + MaxEncodedLen;
    type ProposalNonce: Encode + Decode + Parameter + AtLeast32Bit + Default + Copy + MaxEncodedLen;
    type MaintainerNonce: Encode + Decode + Parameter + AtLeast32Bit + Default + Copy + MaxEncodedLen;
    type SignatureVerifier: SigningSystem;
    type ChainIdentifier: Get<Self::ChainId>;
    type ChainType: Get<[u8; 2]>;
    type ProposalLifetime: Get<Self::BlockNumber>;
    type BridgeAccountId: Get<PalletId>;
    type MaxStringLength: Get<u32>;
    type WeightInfo: WeightInfo;

}

The module configuration trait.

Required Associated Types

type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type AdminOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin used to administer the pallet

type Proposal: Parameter + Dispatchable<RuntimeOrigin = Self::RuntimeOrigin> + EncodeLike + Decode + GetDispatchInfo

Proposed dispatchable call

type SetResourceProposalFilter: Contains<Self::Proposal>

Call filter for proposals

type ExecuteProposalFilter: Contains<Self::Proposal>

type ChainId: Encode + Decode + Parameter + AtLeast32Bit + Default + Copy + From<u64> + From<u32> + MaxEncodedLen

ChainID for anchor edges

type ProposalNonce: Encode + Decode + Parameter + AtLeast32Bit + Default + Copy + MaxEncodedLen

Proposal nonce type

type MaintainerNonce: Encode + Decode + Parameter + AtLeast32Bit + Default + Copy + MaxEncodedLen

Maintainer nonce type

type SignatureVerifier: SigningSystem

Signature verification utility over public key infrastructure

type ChainIdentifier: Get<Self::ChainId>

The identifier for this chain. This must be unique and must not collide with existing IDs within a set of bridged chains.

type ChainType: Get<[u8; 2]>

The chain type for this chain. This is either a standalone Substrate chain, relay chain, or parachain

type ProposalLifetime: Get<Self::BlockNumber>

type BridgeAccountId: Get<PalletId>

type MaxStringLength: Get<u32>

type WeightInfo: WeightInfo